+++ src/main/scala/codegen/ClojureCodeGenerator.scala
@@ -179,7 +179,7 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
         strBuf ++= str
 
       case FStringLiteral(str) =>
-        strBuf ++= s"(fmt ${str})"
+        strBuf ++= s"(fmt $str)"
 
       case DateTimeLiteral(date) =>
         strBuf ++= "#inst  \"" + s"$date" + "\""
@@ -188,7 +188,7 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
         strBuf ++= value
 
       case RegexpLiteral(re) =>
-        strBuf ++= "#\"" + re.stripPrefix("/").stripSuffix("/") + "\""
+        strBuf ++= "#\"" + re + "\""
 
       case MatchExpr(expr, re) =>
         strBuf ++= s"(some? (re-matches ${toClojure(re)} ${toClojure(expr)}))"
@@ -469,14 +469,15 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
                   letDecls = s"[$head & $tail] ${namedArgs.head}" :: letDecls
                 case DefArg(ConsExpression(Identifier(head), ConsExpression(Identifier(second), Identifier(tail)))) =>
                   letDecls = s"[$head $second & $tail] ${namedArgs.head}" :: letDecls
-                case DefArg(ListExpression(args, None)) =>
-                  letDecls = s"[${args.map(toClojure).mkString(" ")}] ${namedArgs.head}]" :: letDecls
-                case DefArg(ConstructorExpression(_, cls, args)) =>
+                case DefArg(ListExpression(defArgs, None)) =>
+                  letDecls = s"[${defArgs.map(toClojure).mkString(" ")}] ${namedArgs.head}]" :: letDecls
+
+                case DefArg(ConstructorExpression(_, cls, ctorArgs)) =>
                   andList = s"(isa-type? $cls ${namedArgs.head})" :: andList
                   var argDecls = List.empty[String]
-                  for (arg <- args) {
+                  for (arg <- ctorArgs) {
                     arg match {
-                      case Identifier(id) => argDecls = s"${id} (.${id} ${namedArgs.head})" :: argDecls
+                      case Identifier(id) => argDecls = s"$id (.$id ${namedArgs.head})" :: argDecls
                     }
                   }
                   letDecls = argDecls.reverse ++ letDecls
@@ -556,7 +557,7 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
 
   def toClojureLetId(id: LetId) : String = {
     id match {
-      case LetSimpleId(id) => id
+      case LetSimpleId(name) => name
       case LetTupledId(ids) => s"[${ids.map(toClojureLetId).mkString(" ")}]"
     }
   }
@@ -621,15 +622,8 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
 
   def toClojureDefMatchArg(defArg: DefArg): String = {
     defArg match {
-      case DefOtherwiseArg => ":default"
-      case DefArg(Identifier(id)) => id
-      case DefArg(TupleExpr(exprs)) => s"[${exprs.map(toClojure).mkString(" ")}]"
-      case DefArg(InfiniteTupleExpr(exprs)) =>
-        val rest = exprs.last
-        val args = exprs.dropRight(1)
-        s"[${args.map(toClojure).mkString(" ")} & ${toClojure(rest)}]"
       case DefArg(ConstructorExpression(_, cls, _)) => s"$cls"
-      case DefArg(expression) => s"${toClojure(expression)}"
+      case d => toClojureDefArg(d)
     }
   }
 
@@ -703,12 +697,11 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
 
   def toClojureImportClauses(importClauses: List[ImportClause]): String = {
     val strBuf = new StringBuilder()
-    val (cljImp, jvmImp) = importClauses.span(p =>
-      p match {
-        case CljImport(_) => true
-        case FromCljRequire(_, _) => true
-        case _ => false
-      })
+    val (cljImp, jvmImp) = importClauses.span {
+      case CljImport(_) => true
+      case FromCljRequire(_, _) => true
+      case _ => false
+    }
     if (cljImp.nonEmpty) {
       strBuf ++= s"(:require [${cljImp.map(toClojureImportClause).mkString(" ")}]) "
     }
@@ -737,7 +730,6 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
     importClause match {
       case JvmImport(name) => name.map(toClojureImportAlias).mkString(" ")
       case FromJvmRequire(from, names) =>
-        val renames = names.filter(p => p.alias.nonEmpty)
         s"($from ${names.map(n => n.name).mkString(" ")}) "
       case _ => ""
     }
@@ -752,7 +744,7 @@ class ClojureCodeGenerator(node: LangNode) extends CodeGenerator {
 
   def toClojureImporteRename(importAlias: ImportAlias) : String = {
     importAlias match {
-      case ImportAlias(name, None) => ""
+      case ImportAlias(_, None) => ""
       case ImportAlias(name, Some(alias)) => s"$name $alias"
     }
   }